add security nonce

+++ admin/general-settings.php
@@ -15,14 +15,14 @@
  * @since    1.0.0
  */
 function ilove_pdf_register_action() {
-    if ( isset( $_POST['action'] ) && 'ilovepdf_register' === $_POST['action'] ) {
+    if ( isset( $_POST['action'] ) && 'ilovepdf_register' === $_POST['action'] && isset( $_POST['nonce_ilove_pdf_register'] ) && wp_verify_nonce( sanitize_key( $_POST['nonce_ilove_pdf_register'] ), 'admin-post' ) ) {
 
         if ( get_option( 'ilovepdf_user_register_times' ) !== null ) {
             $response = wp_remote_post(
                 ILOVE_PDF_REGISTER_URL,
                 array(
 					'body' => array(
-						'name'         => sanitize_text_field( $_POST['ilove_pdf_account_name'] ),
+						'name'         => sanitize_text_field( wp_unslash( $_POST['ilove_pdf_account_name'] ) ),
 						'email'        => sanitize_email( $_POST['ilove_pdf_account_email'] ),
 						'new_password' => sanitize_text_field( $_POST['ilove_pdf_account_password'] ),
 						'free_files'   => 0,
@@ -54,7 +54,7 @@ function ilove_pdf_register_action() {
         }
     }
 
-    wp_safe_redirect( wp_get_referer() . '&response_code=' . $response['response']['code'] );
+    wp_safe_redirect( wp_get_referer() . '&response_code=' . $response['response']['code'] . '&nonce_ilove_pdf_response=' . wp_create_nonce() );
 }
 add_action( 'admin_post_ilovepdf_register', 'ilove_pdf_register_action' );
 
@@ -64,7 +64,7 @@ add_action( 'admin_post_ilovepdf_register', 'ilove_pdf_register_action' );
  * @since    1.0.0
  */
 function ilove_pdf_login_action() {
-    if ( isset( $_POST['action'] ) && 'ilovepdf_login' === $_POST['action'] ) {
+    if ( isset( $_POST['action'] ) && 'ilovepdf_login' === $_POST['action'] && wp_verify_nonce( $_POST['nonce_ilove_pdf_login'], 'admin-post' ) ) {
         $response = wp_remote_post(
             ILOVE_PDF_LOGIN_URL,
             array(
@@ -85,7 +85,8 @@ function ilove_pdf_login_action() {
             add_option( 'ilovepdf_user_id', $user['id'] );
         }
     }
-    wp_safe_redirect( wp_get_referer() . '&response_code=' . $response['response']['code'] );
+
+    wp_safe_redirect( wp_get_referer() . '&response_code=' . $response['response']['code'] . '&nonce_ilove_pdf_response=' . wp_create_nonce() );
 }
 add_action( 'admin_post_ilovepdf_login', 'ilove_pdf_login_action' );
 
@@ -95,7 +96,7 @@ add_action( 'admin_post_ilovepdf_login', 'ilove_pdf_login_action' );
  * @since    1.0.0
  */
 function ilove_pdf_logout_action() {
-    if ( isset( $_GET['action'] ) && 'ilovepdf_logout' === $_GET['action'] ) {
+    if ( isset( $_GET['action'] ) && 'ilovepdf_logout' === $_GET['action'] && isset( $_GET['nonce_ilove_pdf_logout'] ) && wp_verify_nonce( sanitize_key( $_GET['nonce_ilove_pdf_logout'] ), 'admin-post' ) ) {
         delete_option( 'ilovepdf_user_token' );
         delete_option( 'ilovepdf_user_email' );
         delete_option( 'ilovepdf_user_private_key' );
@@ -113,7 +114,7 @@ add_action( 'admin_post_ilovepdf_logout', 'ilove_pdf_logout_action' );
  * @since    1.0.0
  */
 function ilove_pdf_change_project_action() {
-    if ( isset( $_GET['action'] ) && 'ilovepdf_change_project' === $_GET['action'] ) {
+    if ( isset( $_GET['action'] ) && 'ilovepdf_change_project' === $_GET['action'] && isset( $_POST['nonce_ilove_pdf_change_project'] ) && wp_verify_nonce( sanitize_key( $_POST['nonce_ilove_pdf_change_project'] ), 'admin-post' ) ) {
         $stats = ilove_pdf_get_statistics();
         update_option( 'ilovepdf_user_private_key', $stats['projects'][ sanitize_text_field( $_POST['ilovepdf_select_project'] ) ]['secret_key'] );
         update_option( 'ilovepdf_user_public_key', $stats['projects'][ sanitize_text_field( $_POST['ilovepdf_select_project'] ) ]['public_key'] );
